pyproject.toml
[project]
name = "qrapp-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "beanie>=1.30.0",
    "cryptography>=45.0.3",
    "fastapi[standard]>=0.115.12",
    "loguru>=0.7.3",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.1",
    "redis>=6.2.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]
